sidear click function done
