+++ components/index.ts
@@ -9,4 +9,4 @@ export * from './features/dashboard/components';
 export * from './features/account/components';
 
 // Marketing components
-export * from './marketing'; 
+// export * from './marketing';
