+++ components/shared/layout/TopNav.tsx
@@ -1,8 +1,25 @@
+"use client";
+
 import React from "react";
 import Image from "next/image";
 import Searchbar from "../common/Searchbar";
-import { Notification } from "../ui/icons/Notification";
-import { Dropdown } from "../ui/icons";
+import { useSidebar } from "@/context/SidebarContext";
+import { Menu } from "lucide-react";
+
+export const SidebarToggle = () => {
+  const { toggleSidebar } = useSidebar();
+
+  return (
+    <button
+      onClick={toggleSidebar}
+      className="p-2 rounded-md hover:bg-gray-100 dark:hover:bg-gray-800"
+      aria-label="Toggle sidebar"
+    >
+      <Menu className="h-6 w-6" />
+    </button>
+  );
+};
+
 
 const TopNav = () => {
   return (
@@ -13,89 +30,40 @@ const TopNav = () => {
       </div>
 
       {/* Right Side Controls */}
-
-      <div className="flex flex-wrap sm:flex-nowrap items-center w-full sm:w-auto gap-3">
+      <div className="hidden md:flex flex-wrap sm:flex-nowrap items-center w-full sm:w-auto gap-3">
         {/* Network Selector */}
-        <div
-          className="flex items-center text-white text-sm"
-          style={{
-            width: "135px",
-            height: "43.78125px",
-            gap: "8px",
-            paddingTop: "8px",
-            paddingRight: "12px",
-            paddingBottom: "8px",
-            paddingLeft: "12px",
-            borderRadius: "33.2px",
-            border: "0.33px solid #FFFFFF",
-            background: "#FFFFFF1C",
-          }}
-        >
+        <button className="flex cursor-pointer hover:bg-white/30 items-center text-white text-sm justify-between border py-2 px-4 py- w-[139px] rounded-full">
           <Image
             src="/icons/stellar.png"
             alt="Stellar"
-            width={16}
-            height={16}
+            width={22}
+            height={22}
           />
           <span>Stellar</span>
-          <svg
-            className="w-3 h-3 text-white"
-            width="10"
-            height="6"
-            viewBox="0 0 10 6"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
+          <svg className="w-3 h-3 text-white" viewBox="0 0 10 6" fill="none">
             <path
               d="M5 6.0006L0.757324 1.758L2.17154 0.34375L5 3.1722L7.8284 0.34375L9.2426 1.758L5 6.0006Z"
               fill="#FFFFFF"
             />
           </svg>
-        </div>
+        </button>
 
         {/* Wallet Address */}
-        <div
-          className="flex items-center text-white text-sm"
-          style={{
-            width: "139px",
-            height: "43.78125px",
-            gap: "8px",
-            paddingTop: "8px",
-            paddingRight: "12px",
-            paddingBottom: "8px",
-            paddingLeft: "12px",
-            borderRadius: "33.2px",
-            border: "0.33px solid #FFFFFF",
-            background: "#FFFFFF1C",
-          }}
-        >
+        <button className="flex cursor-pointer hover:bg-white/30 items-center text-white text-sm justify-between border py-2 px-4 py- w-[139px] rounded-full">
           <span>Ga2j6...f5g3</span>
-          <svg
-            className="w-3 h-3 text-white"
-            width="10"
-            height="6"
-            viewBox="0 0 10 6"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
+          <svg className="w-3 h-3 text-white" viewBox="0 0 10 6" fill="none">
             <path
               d="M5 6.0006L0.757324 1.758L2.17154 0.34375L5 3.1722L7.8284 0.34375L9.2426 1.758L5 6.0006Z"
               fill="#FFFFFF"
             />
           </svg>
-        </div>
+        </button>
 
         {/* Divider */}
-        <div
-          className="h-8 border-l"
-          style={{
-            borderWidth: "1px",
-            borderColor: "#71B48D",
-          }}
-        ></div>
+        <div className="h-8 border-l" style={{ borderColor: "#71B48D" }}></div>
 
-        {/* Notification Icon */}
-        <div className=" p-2 rounded cursor-pointer">
+        {/* Notification */}
+        <div className="p-2 rounded cursor-pointer">
           <Image
             src="/icons/notification.png"
             alt="Notifications"
@@ -112,6 +80,11 @@ const TopNav = () => {
           width={32}
           height={32}
         />
+      </div>
+
+      {/* ☰ Sidebar Toggle */}
+      <div className="md:hidden">
+        <SidebarToggle />
       </div>
     </div>
   );
